+++ src/collective/solr/browser/facets.py
@@ -3,6 +3,7 @@ from operator import itemgetter
 
 import six
 from collective.solr.interfaces import IFacetTitleVocabularyFactory
+from collective.solr.utils import isActive
 from plone.app.layout.viewlets.common import SearchBoxViewlet
 from plone.registry.interfaces import IRegistry
 from Products.CMFPlone.browser.search import Search
@@ -120,22 +121,28 @@ class SearchBox(SearchBoxViewlet, FacetMixin):
 
 class SearchView(Search):
 
-    pass
+    @property
+    def solr_active(self):
+        return isActive()
+
+    def search_facets(self):
+        view = SearchFacetsView(self.context, self.request)
+        return view(results=self.results(batch=False, use_content_listing=False))
 
 
 class SearchFacetsView(BrowserView, FacetMixin):
     """view for displaying facetting info as provided by solr searches"""
 
-    def __call__(self, *args, **kw):
-        self.args = args
-        self.kw = kw
-        return super(SearchFacetsView, self).__call__(*args, **kw)
+    index = ViewPageTemplateFile("templates/facets.pt")
+
+    def __call__(self, results):
+        self.results = results
+        return self.index()
 
     def facets(self):
         """prepare and return facetting info for the given SolrResponse"""
-        results = self.kw.get("results", None)
-        fcs = getattr(results, "facet_counts", None)
-        if results is not None and fcs is not None:
+        fcs = getattr(self.results, "facet_counts", None)
+        if self.results is not None and fcs is not None:
             filter = lambda name, count: name and count > 0
             return convertFacets(fcs.get("facet_fields", {}), self, filter)
         else: