+++ CHANGES.rst
@@ -20,6 +20,8 @@ Changelog
 
 - Fix Classic UI for Plone 6 [gyst]
 
+- Reimplement Solr faceting for Classic UI in Plone 6 [gyst]
+
 9.4.0 (2024-12-14)
 ------------------