+++ CHANGES.rst
@@ -18,6 +18,10 @@ Changelog
 
 - Remove support for Python 3.9, Solr 9.0 and Solr 9.1 [gyst]
 
+- Fix Classic UI for Plone 6 [gyst]
+
+- Reimplement Solr faceting for Classic UI in Plone 6 [gyst]
+
 - Cache Solr download on Github CI actions [gyst]
 
 9.4.0 (2024-12-14)